+++ CHANGELOG.md
@@ -4,6 +4,9 @@ This file is used to list changes made in each version of grafana.
 
 ## Unreleased
 
+- Add more properties to generic oath resource
+- Add warning when extra_options conflict with resource properties
+
 ## 10.2.13 - *2023-04-17*
 
 ## 10.2.12 - *2023-04-07*